Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
GeoAlchemy2==0.14.2
Shapely==2.0.1
Pillow==10.0.0
pytesseract==0.3.10
spacy==3.7.2
transformers==4.33.2
torch==2.0.1
scikit-learn==1.3.0
opencv-python==4.8.0.76
rasterio==1.3.8
geopandas==0.13.2
folium==0.14.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.1
redis==4.6.0
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1
